+++ src/mouse/mouse.rs
@@ -69,6 +69,7 @@ impl Mouse {
         })
     }
 
+    /// The screen associated with the mouse.
     pub fn screen(&self) -> Option<Screen> {
         self.screen.map_or_else(|| None, |screen| Some(Screen::_from(screen, false)))
     }