+++ src/window.rs
@@ -54,8 +54,9 @@ impl HasHandle<WINDOW> for Window {
 }
 
 impl Window {
+    /// The screen associated with the window.
     pub fn screen(&self) -> Option<Screen> {
-        self.screen.map_or_else(|| None, |ptr| Some(Screen::_from(ptr, false)))
+        self.screen.map_or_else(|| None, |screen| Some(Screen::_from(screen, false)))
     }
 }